+++ apx_core/src/crypto/mod.rs
@@ -3,4 +3,5 @@
 pub mod common;
 pub mod ecdsa;
 pub mod eddsa;
+pub mod hashes;
 pub mod rsa;